requirements.txt
openai>=1.12.0
langchain>=0.1.0
langchain-openai>=0.0.5
python-dotenv>=1.0.0
pinecone-client>=3.0.0
tiktoken>=0.5.2
numpy>=1.24.0
pandas>=2.0.0 
httpx>=0.27.0
letta-client